firestore-flask-app/requirements.txt
flask
firebase-admin
google-cloud-firestore
groq
python-dotenv
